fix: separate structured SEO component
